[MECLIPSE-423] Fix NullPointerException when existing Manifest.MF file has no Class-Path element
EclipseManifestWriter and RadManifestWriter are a cut-and-paste. Created AbstractEclipseManifestWriter as shared parent class
and moved all the common stuff up.  Where the code varied EclipseManifestWriter's implementation was chosen.  This was because the variation was often a bug fix.  

The only variation that could not be easily sorted was in addDependencyToClassPath where they only differ at the end:
	EclipseManifestWriter:
		classpath.append( dependency.getEclipseProjectName() + ".jar" );
	RadManifestWriter
		classpath.append( dependency.getArtifactId() + ".jar" );
I have taken the eclipse one as being correct and changed RadPlugin to override getProjectNameForArifact() to return artifact.getArtifactId() to match RadManifestWriter dependency.getArtifactId() as above.  Integration tests still work but I need a RAD person to validate this for me.

Added testProject8 for IT test. Prior to patch this exercised the defect.  After the patch it looks like it is fixed.
However have asked for clarifiction about write() when a new MANIFEST.MF needs to be written but there is already an existing file.
See http://www.nabble.com/m-eclipse-p%3A-EclipseManifestWriter.write%28%29---doesn%27t-retain-any-keys-except-CLASS_PATH-and-MANIFEST_VERSION-td19546126.html
The problem is that the old values will be lost.  Need to check if that is the correct intention.



git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@696565 13f79535-47bb-0310-9956-ffa450edef68