[MECLIPSE-419]
As part of this change a lot of methods on EclipsePlugin were made final or hidden as private.
This is to avoid the problems in the future where subclasses are extending at the wrong points and getting out of sync. e.g. MyEclipse plugin overwrote writeConfiguration() but missed all the WTP changes.

Added javadocs for public/protected methods that were missing docs.

Renamed methods used for subclassing to be consistent with a "Extras" suffix.

MyEclipse now just writes MyEclipse specific files and EclipsePlugin handles the rest.

Updated ITs to include provided jars in classpath.

As part of the EclipsePlugin/WTP handling target/generated-resources/eclipse will be added to the classpath with the generated META-INF/MANIFEST.MF file.

git-svn-id: https://svn.apache.org/repos/asf/maven/plugins/trunk@727284 13f79535-47bb-0310-9956-ffa450edef68